backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
matplotlib
python-multipart
python-dotenv